requirements.txt

plotly
pandas
sqlalchemy
python-dotenv
